# api/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is an API for managing expenses.",
        "title": "You Choose API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "You Choose",
            "url": "http://www.youchoose.com.br",
            "email": "contato@youchoose.com.br"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/lists": {
            "get": {
                "description": "Get a list of movies and votes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lists"
                ],
                "summary": "Get List",
                "parameters": [
                    {
                        "type": "string",
                        "description": "List id",
                        "name": "list_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User id",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/usecases.GetListByIDOutputDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ProblemDetails"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/util.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ProblemDetails"
                        }
                    }
                }
            },
            "post": {
                "description": "Registers a new list in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lists"
                ],
                "summary": "Create a new list",
                "parameters": [
                    {
                        "description": "Movie data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usecases.CreateListInputDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/usecases.CreateListOutputDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ProblemDetails"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/util.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/lists/movies": {
            "post": {
                "description": "Add new movies to list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lists"
                ],
                "summary": "Add movies to list",
                "parameters": [
                    {
                        "description": "AddMoviesList data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usecases.AddMoviesListInputDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/usecases.AddMoviesListOutputDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ProblemDetails"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/util.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/movies": {
            "post": {
                "description": "Registers a new movie in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Movies"
                ],
                "summary": "Create a new movie",
                "parameters": [
                    {
                        "description": "Movie data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usecases.CreateMovieInputDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/usecases.CreateMovieOutputDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ProblemDetails"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/util.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/votes": {
            "post": {
                "description": "Registers a new vote in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Votes"
                ],
                "summary": "Create a new vote",
                "parameters": [
                    {
                        "description": "Vote data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usecases.VoteInputDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/usecases.VoteOutputDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ProblemDetails"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/util.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ProblemDetails"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entities.Combination": {
            "type": "object",
            "properties": {
                "first_movie": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "list_id": {
                    "type": "string"
                },
                "second_movie": {
                    "type": "string"
                }
            }
        },
        "entities.List": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "combinations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Combination"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "deactivated_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "movies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Movie"
                    }
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "entities.Movie": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "deactivated_at": {
                    "type": "string"
                },
                "external_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "poster": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "entities.Vote": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "combination": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deactivated_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "list_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "winner_id": {
                    "type": "string"
                }
            }
        },
        "usecases.AddMoviesListInputDTO": {
            "type": "object",
            "properties": {
                "list_id": {
                    "type": "string"
                },
                "movies": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "usecases.AddMoviesListOutputDTO": {
            "type": "object",
            "properties": {
                "content_message": {
                    "type": "string"
                },
                "success_message": {
                    "type": "string"
                }
            }
        },
        "usecases.CreateListInputDTO": {
            "type": "object",
            "properties": {
                "movies": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "usecases.CreateListOutputDTO": {
            "type": "object",
            "properties": {
                "content_message": {
                    "type": "string"
                },
                "success_message": {
                    "type": "string"
                }
            }
        },
        "usecases.CreateMovieInputDTO": {
            "type": "object",
            "properties": {
                "external_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "poster": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "usecases.CreateMovieOutputDTO": {
            "type": "object",
            "properties": {
                "content_message": {
                    "type": "string"
                },
                "success_message": {
                    "type": "string"
                }
            }
        },
        "usecases.GetListByIDOutputDTO": {
            "type": "object",
            "properties": {
                "list": {
                    "$ref": "#/definitions/entities.List"
                },
                "votes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Vote"
                    }
                }
            }
        },
        "usecases.VoteInputDTO": {
            "type": "object",
            "properties": {
                "combination": {
                    "type": "string"
                },
                "list_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "winner_id": {
                    "type": "string"
                }
            }
        },
        "usecases.VoteOutputDTO": {
            "type": "object",
            "properties": {
                "content_message": {
                    "type": "string"
                },
                "success_message": {
                    "type": "string"
                }
            }
        },
        "util.ProblemDetails": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "instance": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}